+++ frontend/web/src/app/(auth)/sign-up/page.tsx
@@ -8,8 +8,8 @@ import { z } from "zod";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useAuth } from "@/contexts/AuthProviderWithLoading";
-import { LogoComponent } from "@/components/LogoComponent";
-import { ThemeSelector } from "@/components/ThemeSelector";
+import { LogoComponent } from "@/components/utilities/LogoComponent";
+import { ThemeSelector } from "@/components/theme/ThemeSelector";
 import { Eye, EyeOff, User } from "lucide-react";
 import { toast } from "sonner";
