+++ frontend/web/src/app/(dashboard)/admin/educators/page.tsx
@@ -3,8 +3,8 @@
 import React from "react";
 import { useRouter, usePathname } from 'next/navigation';
 import { useState } from "react";
-import EducatorsCardsView_forAdmin from "@/components/EducatorsCardsView_forAdmin";
-import EducatorsList_forAdmin from "@/components/EducatorsList_forAdmin";
+import EducatorsCardsView_forAdmin from "@/components/admin/educators/EducatorsCardsView_forAdmin";
+import EducatorsList_forAdmin from "@/components/admin/educators/EducatorsList_forAdmin";
 import { LayoutGrid, List, UserPlus, User } from "lucide-react";
 
 const Educators = () => {
