+++ frontend/web/src/app/(dashboard)/loading.tsx
@@ -1,4 +1,4 @@
-import { CenteredPageLoader } from '@/components/MinimalisticLoader';
+import { CenteredPageLoader } from '@/components/loaders/MinimalisticLoader';
 
 export default function Loading() {
   return <CenteredPageLoader variant="orbit" message="Loading dashboard..." />;
