+++ frontend/web/src/app/loading.tsx
@@ -1,4 +1,4 @@
-import { CenteredPageLoader } from '@/components/MinimalisticLoader';
+import { CenteredPageLoader } from '@/components/loaders/MinimalisticLoader';
 
 export default function Loading() {
   return <CenteredPageLoader variant="spinner" message="Loading..." />;
