+++ frontend/web/src/app/test-graphql/page.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from 'react';
 import { toast } from 'sonner';
-import { ThemeSelector } from '@/components/ThemeSelector';
+import { ThemeSelector } from '@/components/theme/ThemeSelector';
 
 /**
  * GraphQL Test Interface
