+++ frontend/web/src/components/dashboard-navigation/DashboardNavbar.tsx
@@ -1,7 +1,7 @@
 "use client"
 
-import { ThemeSelector } from "@/components/ThemeSelector"
-import { ProfileDropdown } from "@/components/ProfileDropdown"
+import { ThemeSelector } from "@/components/theme/ThemeSelector"
+import { ProfileDropdown } from "@/components/utilities/ProfileDropdown"
 import { GlassNavbar } from "@/components/glass"
 import { SidebarTrigger } from "@/components/ui/sidebar"
